feat: adjust the number of images per row based on screen size
The number of images per row is fixed, and the resolutions of images on
small window can be too low.

Prevent image from resizing by changing the number of images per row dynamically
based on the current screen size. To get the window size, this change uses ioctl()
system call.