fix: thumbnail smaller than container left aligned
"to_image" method makes the source image into thumbnail and paste
the result to container image. This works fine if the size of image is
greater than or equal to the container size, however, the thumbnail is
left aligned if the source image is smaller.

Adjust the position of a box when pasting a thumbnail to make sure that
it is center aligned.