fix: catch known exceptions, and print error
No exception handling is in place, and even known exceptions bubble up to
the top of the stack. As the result, a user sees exception traceback which can
be not user friendly.

Change the behavior of "core.run" function to take errout IO to print out known
errors before exiting with error code.